fix: JSON parsing
TECHNICALLY this is a change in behaviour, going from catching failed JSON parsing and returning an empty settings object `{}`, to returning the keys with null values instead, even if the object to parse is undefined. Whilst this is a change in behaviour, a grep through the module showed that these are only used as form initialValues, where the existence of the keys should have no bearing, only the values present.

This approach means that the non-existence of ONE set of settings values will no longer prevent the other from showing up in the form, which is better overall behaviour.